move mq specific code to mq-host class
